[Refactor] s3 이미지 업로드 파일 폴더 변경
